expanded the test case to include a check for manually setting the paramactive to [T,T,F]

+++ casatasks/tests/tasks/test_task_fringefit.py
@@ -309,19 +309,27 @@ class Fringefit_paramactive_caltable(unittest.TestCase):
     preapplytable = 'topreapply.cal'
     nocallib = 'nocallib.cal'
     withcallib = 'withcallib.cal'
+    manualdefault = 'manualcallib.cal'
     
     def setUp(self):
         shutil.copytree(os.path.join(datapath, self.msfile), self.msfile)
         
     def tearDown(self):
         shutil.rmtree(self.msfile)
-        shutil.rmtree(self.preapplytable)
-        shutil.rmtree(self.nocallib)
-        shutil.rmtree(self.withcallib)
-        os.remove(self.testcallib)
+        
+        if os.path.exists(self.preapplytable):
+            shutil.rmtree(self.preapplytable)
+        if os.path.exists(self.nocallib):
+            shutil.rmtree(self.nocallib)
+        if os.path.exists(self.withcallib):
+            shutil.rmtree(self.withcallib)
+        if os.path.exists(self.manualdefault):
+            shutil.rmtree(self.manualdefault)
+        if os.path.exists(self.testcallib):
+            os.remove(self.testcallib)
         
     def test_paramactive_callib(self):
-        """ Test that the default state for paramactive with callib is [T,T,F] """
+        """ Test that the default state for paramactive with callib matches [T,T,F]"""
         # create the table to pre-apply
         fringefit(vis=self.msfile, caltable=self.preapplytable, refant='0')
         
@@ -332,6 +340,7 @@ class Fringefit_paramactive_caltable(unittest.TestCase):
         # run with gaintable preapply and with callib and running default paramactive
         fringefit(vis=self.msfile, caltable=self.nocallib, refant='0', docallib=False, gaintable=[self.preapplytable], paramactive=[])
         fringefit(vis=self.msfile, caltable=self.withcallib, refant='0', docallib=True, callib=self.testcallib, paramactive=[])
+        fringefit(vis=self.msfile, caltable=self.manualdefault, refant='0', docallib=True, callib=self.testcallib, paramactive=[True,True,False])
         
         # get the FPARAM data for each table and compare
         tblocal.open(self.nocallib)
@@ -342,9 +351,13 @@ class Fringefit_paramactive_caltable(unittest.TestCase):
         res2 = tblocal.getcol('FPARAM')
         tb.close()
         
+        tblocal.open(self.manualdefault)
+        res3 = tblocal.getcol('FPARAM')
+        tblocal.close()
+        
         self.assertTrue(np.all(res1 == res2), msg='Results differ when preapplying with callib vs gaintable')
+        self.assertTrue(np.all(res2 == res3), msg='Results differ between paramactive [] and [True,True,False]')
     
         
-
 if __name__ == '__main__':
     unittest.main()
